Ballクラスをファイル分割
